Move avatar change from header to profile page
